+++ src/+io/select_input.m
@@ -77,12 +77,12 @@ function [SoilProperties, leafbio, canopy, meteo, angles, SpaceTimeInfo] = selec
 
     %% derived input
     if options.soil_heat_method == 1
-        SoilProperties.GAM =  equations.Soil_Inertia1(SoilProperties.SMC);
+        SoilProperties.GAM =  equations.Soil_Inertia1(SoilProperties.SMC, SoilProperties.theta_s0);
     else
         SoilProperties.GAM  = equations.Soil_Inertia0(SoilProperties.cs, SoilProperties.rhos, SoilProperties.lambdas);
     end
     if options.calc_rss_rbs
-        [SoilProperties.rss, SoilProperties.rbs] = equations.calc_rssrbs(SoilProperties.SMC, canopy.LAI, SoilProperties.rbs);
+        [SoilProperties.rss, SoilProperties.rbs] = equations.calc_rssrbs(SoilProperties, canopy.LAI, SoilProperties.rbs);
     end
 
     if leafbio.Type